requirements.txt
# Core FastAPI and web framework
fastapi==0.135.1
uvicorn[standard]==0.41.0
pydantic[email]==2.12.5
pydantic-settings==2.13.1

# Database and ORM
sqlalchemy==2.0.48
alembic==1.18.4
psycopg2-binary==2.9.11

# Authentication and Security
python-jose[cryptography]==3.5.0
argon2-cffi==25.1.0
python-multipart==0.0.22

# Background Tasks
celery==5.4.0
redis==7.3.0
flower==2.0.1

# Image Processing and Geospatial
rasterio==1.5.0
numpy==2.4.3
scikit-image==0.26.0
opencv-python==4.13.0.92
Pillow==12.1.1
landsatxplore==0.12.1
click==8.3.1

# HTTP Client and APIs
httpx==0.28.1
aiohttp==3.13.3

# Email Service
mailchimp-transactional==1.3.3

# Environment and Configuration
python-decouple==3.8
uuid-v7==1.0.0
supabase==2.28.2

# Development and Debugging
rich==14.3.3

# Testing (for development)
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
faker==40.11.0

# Code Quality
black==26.3.1
isort==8.0.1
flake8==7.3.0

# Documentation
mkdocs==1.6.1
mkdocs-material==9.7.5
